+++ tests/run.R
@@ -36,41 +36,41 @@ local({                                         # so we can use `on.exit`
 
   valgrind <- FALSE
   if(!valgrind) {
-    test.res <- test_dir(
-      "testthat",
-      filter=paste0(                              # so we can run subset of files
-        c(
-          "atomic",
-          "banner",
-          "capture",
-          "check",
-          "context",
-          "core",
-          "diffChr",
-          "diffDeparse",
-          "diffObj",
-          "diffPrint",
-          "diffStr",
-          "file",
-          "guide",
-          "html",
-          "limit",
-          "methods",
-          "misc",
-          if(nchar(Sys.getenv('NOT_CRAN'))) "notcran",
-          "pager",
-          "rdiff",
-          "s4",
-          "ses",     # run this file only for valgrind
-          "style",
-          "subset",
-          "summary",
-          "text",
-          "trim",
-          "warning"
-        ), collapse="|"
-      )
+    filter <- paste0(           # so we can run subset of files
+      c(
+        "atomic",
+        "banner",
+        "capture",
+        "check",
+        "context",
+        "core",
+        "diffChr",
+        "diffDeparse",
+        "diffObj",
+        "diffPrint",
+        "diffStr",
+        "file",
+        "guide",
+        "html",
+        "limit",
+        "methods",
+        "misc",
+        if(nchar(Sys.getenv('NOT_CRAN'))) "notcran",
+        "pager",
+        "rdiff",
+        "s4",
+        "ses",     # run this file only for valgrind
+        "style",
+        "subset",
+        "summary",
+        "text",
+        "trim",
+        "warning"
+      ),
+      collapse="|"
     )
+    test.res <- test_dir("testthat", filter=filter)
+    # test.res <- test_dir("testthat", filter="diffPrint")
     with(
       as.data.frame(test.res), {
         fail <- sum(failed)