+++ tests/testthat/testthat.ses.R
@@ -94,6 +94,10 @@ test_that("corner cases?", {
 test_that("errors", {
   expect_error(ses('a', 'b', max.diffs='hello'), "must be scalar integer")
   expect_error(ses('a', 'b', warn='hello'), "must be TRUE or FALSE")
+
+  a <- structure(1, class='diffobj_ogewlhgiadfl2')
+  expect_error(ses(a, 1), "could not be coerced")
+  expect_error(ses(1, a), "could not be coerced")
 })
 
 # We want to have a test file that fully covers the C code in order to run